Enable cross eventLoop InMemory tests.
